Create README.md
A little about me so that the people that are interested in what I am looking for might have the patience and the know how to teach me so that I can make something of myself. If no one teaches another we will always let Government to control us all the more we fight back with knowledge the more we win as a civilization.